♻️ app.rs から hud_show.rs / config_reload.rs を切り出す
HUD表示とフェード（show_text_content / present_hud / hide_hud 等）を
hud_show.rs に、設定ファイルの再読み込み（display_settings_from_file /
reload_config_if_changed / apply_settings_now）を config_reload.rs に
分離した。extern "C" のうち hide_hud はロック取得だけを mod.rs のシムに
残し実処理を hide_hud_now として切り出したが、fade_tick は AppKit 呼び出し
前に APP_STATE のロックを明示的に手放す構造のため丸ごと hud_show.rs へ
移した。CLAUDE.md のモジュール構成表も src/app/ に更新。
