+++ CLAUDE.md
@@ -46,7 +46,7 @@ UI 文言を README に書くときは `src/i18n.rs` の文言テーブルと一
 | `src/cli.rs` | CLIフラグの処理（`--help`, `--config-show`, `--render-hud-png` など） |
 | `src/config/` | 設定（`types.rs` 型 / `parse.rs` パース / `io.rs` 読み書き / `settings.rs` 項目定義 / `cli.rs` `--config-show` の出力） |
 | `src/hud.rs` | HUDウィンドウ生成・レイアウト計算・描画 |
-| `src/app.rs` | AppDelegate・クリップボード監視・フェードアニメーション |
+| `src/app/` | AppDelegate・クリップボード監視・フェードアニメーション（`mod.rs` デリゲート配線・AppState / `panels.rs` About・支援ページ・自動起動プロンプト / `hud_show.rs` HUD表示とフェード / `config_reload.rs` 設定ファイルの再読み込み） |
 | `src/menu.rs` | メニューバー常駐アイコンとメインメニューの構築 |
 | `src/settings_window/` | 設定ウィンドウのUI（`mod.rs` 型定義・再エクスポート / `rows.rs` コントロール生成・配置 / `build.rs` ウィンドウ組み立て / `sync.rs` AppState との同期・操作） |
 | `src/login_item.rs` | LaunchAgent の plist 書き出し・解除（自動起動） |
